Make -x work properly.  This should close #247330 once and for all.
Refactor spawn_child to move path and shell lookups into their own methods.

Make command a string for -e and a list for -x to distinguish them.

If command is a string, let the shell deal with it, if command is a list,
do a path lookup and run it directly.